Suppress number of files generated to be max_threads in LocalExecutor even if number of input tasks are larger than max_threads.
PURPOSE:

Think of using FileInput plugins such as input-file, input-s3, and there
are a lots of files as inputs. Before, embulk-output-bigquery generated
local files of the same number with input files, and tried to upload them in parallel.

Since we dropped `max_load_parallels` option as a request of frsyuki
at https://github.com/embulk/embulk-output-bigquery/pull/35, number of
concurrency can not be suppressed, which results in consuming lots of
network resources.  Futhermore, loading lots of files into bigquery is
not efficient, we should merge them beforehand.

This feature suppresses number of files to be generated to be same with
max_threads, which results in much more efficient.